refac: repackage admin service and test code
