+++ tests/service/test_user.py
@@ -388,227 +388,6 @@ async def test_exist_by_existing_nickname(
     assert result.is_duplicated is True
 
 
-async def test_approve_center(
-        session: AsyncSession,
-        mock_repo: dict,
-        user_service: UserService,
-        mock_center: Center
-):
-    # given
-    request_user = RequestUser(id="123456", email="test@claon.com", role=Role.ADMIN)
-    center_id = mock_center.id
-
-    mock_repo["center"].find_by_id.side_effect = [mock_center]
-    mock_center.approved = True
-    mock_center.user.role = Role.CENTER_ADMIN
-    mock_repo["center"].approve.side_effect = [mock_center]
-
-    # when
-    result = await user_service.approve_center(center_id, session, request_user)
-
-    # then
-    assert result.approved
-    assert mock_center.user.role == Role.CENTER_ADMIN
-
-
-async def test_approve_center_with_non_admin(
-        session: AsyncSession,
-        mock_repo: dict,
-        user_service: UserService,
-        mock_center: Center
-):
-    # given
-    request_user = RequestUser(id="123456", email="test@claon.com", role=Role.PENDING)
-    center_id = mock_center.id
-
-    # then
-    with pytest.raises(BadRequestException):
-        # when
-        await user_service.approve_center(center_id, session, request_user)
-
-
-async def test_approve_not_existing_center(
-        session: AsyncSession,
-        mock_repo: dict,
-        user_service: UserService
-):
-    # given
-    request_user = RequestUser(id="123456", email="test@claon.com", role=Role.ADMIN)
-    center_id = "not_existing_id"
-
-    mock_repo["center"].find_by_id.side_effect = [None]
-
-    # then
-    with pytest.raises(BadRequestException):
-        # when
-        await user_service.approve_center(center_id, session, request_user)
-
-
-async def test_approve_lector(
-        session: AsyncSession,
-        mock_repo: dict,
-        user_service: UserService,
-        mock_lector: Lector
-):
-    # given
-    request_user = RequestUser(id="123456", email="test@claon.com", role=Role.ADMIN)
-    lector_id = mock_lector.id
-
-    mock_repo["lector"].find_by_id.side_effect = [mock_lector]
-    mock_lector.approved = True
-    mock_lector.user.role = Role.LECTOR
-    mock_repo["lector"].approve.side_effect = [mock_lector]
-
-    # when
-    result = await user_service.approve_lector(lector_id, session, request_user)
-
-    # then
-    assert result.approved
-    assert mock_lector.user.role == Role.LECTOR
-
-
-async def test_approve_lector_with_non_admin(
-        session: AsyncSession,
-        mock_repo: dict,
-        user_service: UserService,
-        mock_lector: Lector
-):
-    # given
-    request_user = RequestUser(id="123456", email="test@claon.com", role=Role.PENDING)
-    lector_id = mock_lector.id
-
-    # then
-    with pytest.raises(BadRequestException):
-        # when
-        await user_service.approve_lector(lector_id, session, request_user)
-
-
-async def test_approve_not_existing_lector(
-        session: AsyncSession,
-        mock_repo: dict,
-        user_service: UserService,
-        mock_lector: Lector
-):
-    # given
-    request_user = RequestUser(id="123456", email="test@claon.com", role=Role.ADMIN)
-    lector_id = "not_existing_id"
-
-    mock_repo["lector"].find_by_id.side_effect = [None]
-
-    # then
-    with pytest.raises(BadRequestException):
-        # when
-        await user_service.approve_lector(lector_id, session, request_user)
-
-
-async def test_reject_center(
-        session: AsyncSession,
-        mock_repo: dict,
-        user_service: UserService,
-        mock_center: Center,
-        mock_center_approved_files: CenterApprovedFile
-):
-    # given
-    request_user = RequestUser(id="123456", email="test@claon.com", role=Role.ADMIN)
-    center_id = mock_center.id
-
-    mock_repo["center"].find_by_id.side_effect = [mock_center]
-
-    # when
-    result = await user_service.reject_center(center_id, session, request_user)
-
-    # then
-    assert result is None
-
-
-async def test_reject_center_with_non_admin(
-        session: AsyncSession,
-        mock_repo: dict,
-        user_service: UserService,
-        mock_center: Center
-):
-    # given
-    request_user = RequestUser(id="123456", email="test@claon.com", role=Role.PENDING)
-    center_id = mock_center.id
-
-    # then
-    with pytest.raises(BadRequestException):
-        # when
-        await user_service.reject_center(center_id, session, request_user)
-
-
-async def test_reject_not_existing_center(
-        session: AsyncSession,
-        mock_repo: dict,
-        user_service: UserService,
-        mock_center: Center
-):
-    # given
-    request_user = RequestUser(id="123456", email="test@claon.com", role=Role.PENDING)
-    center_id = "not_existing_id"
-
-    mock_repo["center"].find_by_id.side_effect = [None]
-
-    # then
-    with pytest.raises(BadRequestException):
-        # when
-        await user_service.reject_center(center_id, session, request_user)
-
-
-async def test_reject_lector(
-        session: AsyncSession,
-        mock_repo: dict,
-        user_service: UserService,
-        mock_lector: Lector,
-        mock_lector_approved_files: LectorApprovedFile
-):
-    # given
-    request_user = RequestUser(id="123456", email="test@claon.com", role=Role.ADMIN)
-    lector_id = mock_lector.id
-
-    mock_repo["lector"].find_by_id.side_effect = [mock_lector]
-    mock_repo["lector"].delete.side_effect = [mock_lector]
-
-    # when
-    result = await user_service.reject_lector(lector_id, session, request_user)
-
-    # then
-    assert result is mock_lector
-
-
-async def test_reject_lector_with_non_admin(
-        session: AsyncSession,
-        mock_repo: dict,
-        user_service: UserService,
-        mock_lector: Lector
-):
-    # given
-    request_user = RequestUser(id="123456", email="test@claon.com", role=Role.PENDING)
-    lector_id = mock_lector.id
-
-    # then
-    with pytest.raises(BadRequestException):
-        # when
-        await user_service.reject_lector(lector_id, session, request_user)
-
-
-async def test_reject_not_existing_lector(
-        session: AsyncSession,
-        mock_repo: dict,
-        user_service: UserService,
-        mock_lector: Lector
-):
-    # given
-    request_user = RequestUser(id="123456", email="test@claon.com", role=Role.PENDING)
-    lector_id = "not_existing_id"
-
-    mock_repo["lector"].find_by_id.side_effect = [None]
-
-    # then
-    with pytest.raises(BadRequestException):
-        # when
-        await user_service.reject_lector(lector_id, session, request_user)
-
 @pytest.mark.asyncio
 @patch("claon_admin.service.user.create_refresh_token")
 async def test_sign_in_with_google(
